refactor(conversation-transformer): use file templates for JS resolvers (#2897)
