+++ packages/amplify-graphql-conversation-transformer/src/resolvers/invoke-lambda-resolver.ts
@@ -1,7 +1,9 @@
 import { TransformerContextProvider, MappingTemplateProvider } from '@aws-amplify/graphql-transformer-interfaces';
 import { MappingTemplate } from '@aws-amplify/graphql-transformer-core';
 import { ConversationDirectiveConfiguration } from '../grapqhl-conversation-transformer';
-import { dedent } from 'ts-dedent';
+import fs from 'fs';
+import path from 'path';
+import dedent from 'ts-dedent';
 
 /**
  * Creates a mapping template for invoking a Lambda function in the context of a GraphQL conversation.
@@ -14,27 +16,42 @@ export const invokeLambdaMappingTemplate = (
   config: ConversationDirectiveConfiguration,
   ctx: TransformerContextProvider,
 ): MappingTemplateProvider => {
-  const req = createInvokeLambdaRequestFunction(config, ctx);
-  const res = createInvokeLambdaResponseFunction(config);
-  return MappingTemplate.inlineTemplateFromString(dedent(req + '\n' + res));
+  const { TOOL_DEFINITIONS_LINE, TOOLS_CONFIGURATION_LINE } = generateToolLines(config);
+  const SELECTION_SET = selectionSet;
+  const GRAPHQL_API_ENDPOINT = ctx.api.graphqlUrl;
+  const MODEL_CONFIGURATION_LINE = generateModelConfigurationLine(config);
+  const RESPONSE_MUTATION_NAME = config.responseMutationName;
+  const RESPONSE_MUTATION_INPUT_TYPE_NAME = config.responseMutationInputTypeName;
+  const MESSAGE_MODEL_NAME = config.messageModel.messageModel.name.value;
+
+  const substitutions = {
+    TOOL_DEFINITIONS_LINE,
+    TOOLS_CONFIGURATION_LINE,
+    SELECTION_SET,
+    GRAPHQL_API_ENDPOINT,
+    MODEL_CONFIGURATION_LINE,
+    RESPONSE_MUTATION_NAME,
+    RESPONSE_MUTATION_INPUT_TYPE_NAME,
+    MESSAGE_MODEL_NAME,
+  };
+
+  let resolver = fs.readFileSync(path.join(__dirname, 'invoke-lambda-resolver-fn.template.js'), 'utf8');
+  Object.entries(substitutions).forEach(([key, value]) => {
+    const replaced = resolver.replace(new RegExp(`\\[\\[${key}\\]\\]`, 'g'), value);
+    resolver = replaced;
+  });
+
+  // This unfortunately needs to be an inline template because an s3 mapping template doesn't allow the CDK
+  // to substitute token values, which is necessary for this resolver function due to its reference of
+  // `ctx.api.graphqlUrl`.
+  return MappingTemplate.inlineTemplateFromString(resolver);
 };
 
-/**
- * Creates a request function for invoking a Lambda function in the context of a GraphQL conversation.
- * This function prepares the necessary data and configuration for the Lambda invocation.
- *
- * @param {ConversationDirectiveConfiguration} config - The configuration for the conversation directive.
- * @param {TransformerContextProvider} ctx - The transformer context provider.
- * @returns {MappingTemplateProvider} A function that generates the request mapping template.
- */
-const createInvokeLambdaRequestFunction = (config: ConversationDirectiveConfiguration, ctx: TransformerContextProvider): string => {
-  const { responseMutationInputTypeName, responseMutationName } = config;
+const generateToolLines = (config: ConversationDirectiveConfiguration) => {
   const toolDefinitions = JSON.stringify(config.toolSpec);
-  const toolDefinitionsLine = toolDefinitions ? `const toolDefinitions = ${toolDefinitions};` : '';
-  const modelConfigurationLine = generateModelConfigurationLine(config);
-  const graphqlEndpoint = ctx.api.graphqlUrl;
+  const TOOL_DEFINITIONS_LINE = toolDefinitions ? `const toolDefinitions = ${toolDefinitions};` : '';
 
-  const toolsConfigurationLine = toolDefinitions
+  const TOOLS_CONFIGURATION_LINE = toolDefinitions
     ? dedent`const dataTools = toolDefinitions.tools;
      const toolsConfiguration = {
       dataTools,
@@ -44,70 +61,7 @@ const createInvokeLambdaRequestFunction = (config: ConversationDirectiveConfigur
       clientTools
     };`;
 
-  const requestFunctionString = `
-  import { util } from '@aws-appsync/utils';
-
-  export function request(ctx) {
-    const { args, identity, request, prev } = ctx;
-    ${toolDefinitionsLine}
-    const selectionSet = '${selectionSet}';
-    const graphqlApiEndpoint = '${graphqlEndpoint}';
-
-    const messages = prev.result.items;
-    const responseMutation = {
-      name: '${responseMutationName}',
-      inputTypeName: '${responseMutationInputTypeName}',
-      selectionSet,
-    };
-    const currentMessageId = ctx.stash.defaultValues.id;
-    ${modelConfigurationLine}
-
-    const clientTools = args.toolConfiguration?.tools?.map((tool) => { return { ...tool.toolSpec }});
-    ${toolsConfigurationLine}
-
-    const authHeader = request.headers['authorization'];
-    const payload = {
-      conversationId: args.conversationId,
-      currentMessageId,
-      responseMutation,
-      graphqlApiEndpoint,
-      modelConfiguration,
-      request: { headers: { authorization: authHeader }},
-      messages,
-      toolsConfiguration,
-    };
-
-    return {
-      operation: 'Invoke',
-      payload,
-      invocationType: 'Event'
-    };
-  }`;
-
-  return requestFunctionString;
-};
-
-const createInvokeLambdaResponseFunction = (config: ConversationDirectiveConfiguration): string => {
-  const responseFunctionString = `
-export function response(ctx) {
-  let success = true;
-  if (ctx.error) {
-    util.appendError(ctx.error.message, ctx.error.type);
-    success = false;
-  }
-  const response = {
-      __typename: '${config.messageModel.messageModel.name.value}',
-      id: ctx.stash.defaultValues.id,
-      conversationId: ctx.args.conversationId,
-      role: 'user',
-      content: ctx.args.content,
-      createdAt: ctx.stash.defaultValues.createdAt,
-      updatedAt: ctx.stash.defaultValues.updatedAt,
-  };
-  return response;
-}`;
-
-  return responseFunctionString;
+  return { TOOL_DEFINITIONS_LINE, TOOLS_CONFIGURATION_LINE };
 };
 
 /**
