+++ packages/amplify-graphql-conversation-transformer/src/transformer-steps/conversation-resolver-generator.ts
@@ -48,14 +48,22 @@ export class ConversationResolverGenerator {
     const { functionDataSourceId, referencedFunction } = this.setupFunctionDataSource(directive, functionStack, capitalizedFieldName);
 
     this.createAssistantResponseResolver(ctx, directive, capitalizedFieldName);
-    this.createAssistantResponseSubscriptionResolver(ctx, capitalizedFieldName);
+    this.createAssistantResponseSubscriptionResolver(ctx, directive, capitalizedFieldName);
 
     const functionDataSource = this.addLambdaDataSource(ctx, functionDataSourceId, referencedFunction, capitalizedFieldName);
     const invokeLambdaFunction = invokeLambdaMappingTemplate(directive, ctx);
 
     this.setupMessageTableIndex(ctx, directive);
 
-    this.createConversationPipelineResolver(ctx, parentName, fieldName, capitalizedFieldName, functionDataSource, invokeLambdaFunction);
+    this.createConversationPipelineResolver(
+      ctx,
+      parentName,
+      fieldName,
+      capitalizedFieldName,
+      functionDataSource,
+      invokeLambdaFunction,
+      directive,
+    );
   }
 
   /**
@@ -161,23 +169,22 @@ export class ConversationResolverGenerator {
     capitalizedFieldName: string,
     functionDataSource: any,
     invokeLambdaFunction: MappingTemplateProvider,
+    directive: ConversationDirectiveConfiguration,
   ): void {
     const resolverResourceId = ResolverResourceIDs.ResolverResourceID(parentName, fieldName);
-    const mappingTemplate = {
-      codeMappingTemplate: invokeLambdaFunction,
-    };
+    const runtime = APPSYNC_JS_RUNTIME;
     const conversationPipelineResolver = new TransformerResolver(
       parentName,
       fieldName,
       resolverResourceId,
-      mappingTemplate,
+      { codeMappingTemplate: invokeLambdaFunction },
       ['init', 'auth', 'verifySessionOwner', 'writeMessageToTable', 'retrieveMessageHistory'],
       ['handleLambdaResponse', 'finish'],
       functionDataSource,
-      APPSYNC_JS_RUNTIME,
+      runtime,
     );
 
-    this.addPipelineResolverFunctions(ctx, conversationPipelineResolver, capitalizedFieldName);
+    this.addPipelineResolverFunctions(ctx, conversationPipelineResolver, capitalizedFieldName, directive);
 
     ctx.resolvers.addResolver(parentName, fieldName, conversationPipelineResolver);
   }
@@ -189,17 +196,22 @@ export class ConversationResolverGenerator {
    * @param capitalizedFieldName - The capitalized field name
    * @param runtime - The runtime configuration
    */
-  private addPipelineResolverFunctions(ctx: TransformerContextProvider, resolver: TransformerResolver, capitalizedFieldName: string): void {
+  private addPipelineResolverFunctions(
+    ctx: TransformerContextProvider,
+    resolver: TransformerResolver,
+    capitalizedFieldName: string,
+    directive: ConversationDirectiveConfiguration,
+  ): void {
     // Add init function
-    const initFunction = initMappingTemplate();
+    const initFunction = initMappingTemplate(directive);
     resolver.addJsFunctionToSlot('init', initFunction);
 
     // Add auth function
-    const authFunction = authMappingTemplate();
+    const authFunction = authMappingTemplate(directive);
     resolver.addJsFunctionToSlot('auth', authFunction);
 
     // Add verifySessionOwner function
-    const verifySessionOwnerFunction = verifySessionOwnerMappingTemplate();
+    const verifySessionOwnerFunction = verifySessionOwnerMappingTemplate(directive);
     const sessionModelName = `Conversation${capitalizedFieldName}`;
     const sessionModelDDBDataSourceName = getModelDataSourceNameForTypeName(ctx, sessionModelName);
     const conversationSessionDDBDataSource = ctx.api.host.getDataSource(sessionModelDDBDataSourceName);
@@ -213,7 +225,7 @@ export class ConversationResolverGenerator {
     resolver.addJsFunctionToSlot('writeMessageToTable', writeMessageToTableFunction, messageDDBDataSource as any);
 
     // Add retrieveMessageHistory function
-    const retrieveMessageHistoryFunction = readHistoryMappingTemplate();
+    const retrieveMessageHistoryFunction = readHistoryMappingTemplate(directive);
     resolver.addJsFunctionToSlot('retrieveMessageHistory', retrieveMessageHistoryFunction, messageDDBDataSource as any);
   }
 
@@ -229,18 +241,14 @@ export class ConversationResolverGenerator {
     capitalizedFieldName: string,
   ): void {
     const assistantResponseResolverResourceId = ResolverResourceIDs.ResolverResourceID('Mutation', directive.responseMutationName);
-    const assistantResponseResolverFunction = assistantMutationResolver();
+    const assistantResponseResolverFunction = assistantMutationResolver(directive);
     const conversationMessageDataSourceName = getModelDataSourceNameForTypeName(ctx, `ConversationMessage${capitalizedFieldName}`);
     const conversationMessageDataSource = ctx.api.host.getDataSource(conversationMessageDataSourceName);
-
-    const mappingTemplate = {
-      codeMappingTemplate: assistantResponseResolverFunction,
-    };
     const assistantResponseResolver = new TransformerResolver(
       'Mutation',
       directive.responseMutationName,
       assistantResponseResolverResourceId,
-      mappingTemplate,
+      { codeMappingTemplate: assistantResponseResolverFunction },
       [],
       [],
       conversationMessageDataSource as any,
@@ -255,13 +263,17 @@ export class ConversationResolverGenerator {
    * @param ctx - The transformer context provider
    * @param capitalizedFieldName - The capitalized field name
    */
-  private createAssistantResponseSubscriptionResolver(ctx: TransformerContextProvider, capitalizedFieldName: string): void {
+  private createAssistantResponseSubscriptionResolver(
+    ctx: TransformerContextProvider,
+    directive: ConversationDirectiveConfiguration,
+    capitalizedFieldName: string,
+  ): void {
     const onAssistantResponseSubscriptionFieldName = `onCreateAssistantResponse${capitalizedFieldName}`;
     const onAssistantResponseSubscriptionResolverResourceId = ResolverResourceIDs.ResolverResourceID(
       'Subscription',
       onAssistantResponseSubscriptionFieldName,
     );
-    const onAssistantResponseSubscriptionResolverFunction = conversationMessageSubscriptionMappingTamplate();
+    const onAssistantResponseSubscriptionResolverFunction = conversationMessageSubscriptionMappingTamplate(directive);
 
     const mappingTemplate = {
       codeMappingTemplate: onAssistantResponseSubscriptionResolverFunction,
